Cart, toast done. Next: order and order history

+++ backend/app/modules/cart/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, String, Integer, Numeric, Boolean, ForeignKey, Text, Enum, Date, Float
+from sqlalchemy import BigInteger, Column, DateTime, String, Integer, Numeric, Boolean, ForeignKey, Text, Enum, Date, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -6,7 +6,7 @@ Base = declarative_base()
 
 class CartItems(Base):
     __tablename__ = "cart_items"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     user_id = Column(String(100), nullable=False) #logto_id
     product_id = Column(String(50), nullable=False)
     quantity = Column(Integer, default=1)
